+++ tests/src/HighlightController.cpp
@@ -11,7 +11,6 @@
 #include "singletons/Paths.hpp"
 #include "singletons/Settings.hpp"
 
-#include <boost/optional/optional_io.hpp>
 #include <gmock/gmock.h>
 #include <gtest/gtest.h>
 #include <QDebug>
@@ -268,7 +267,7 @@ TEST_F(HighlightControllerTest, LoggedInAndConfigured)
                 {
                     false,                                  // alert
                     false,                                  // playsound
-                    boost::none,                            // custom sound url
+                    std::nullopt,                           // custom sound url
                     std::make_shared<QColor>("#7fffffff"),  // color
                     false,
                 },
@@ -307,7 +306,7 @@ TEST_F(HighlightControllerTest, LoggedInAndConfigured)
                 {
                     true,                                   // alert
                     false,                                  // playsound
-                    boost::none,                            // custom sound url
+                    std::nullopt,                           // custom sound url
                     std::make_shared<QColor>("#7fe8b7eb"),  // color
                     false,                                  //showInMentions
                 },
@@ -332,7 +331,7 @@ TEST_F(HighlightControllerTest, LoggedInAndConfigured)
                 {
                     true,                                   // alert
                     false,                                  // playsound
-                    boost::none,                            // custom sound url
+                    std::nullopt,                           // custom sound url
                     std::make_shared<QColor>("#7fffffff"),  // color
                     false,                                  //showInMentions
                 },
@@ -358,7 +357,7 @@ TEST_F(HighlightControllerTest, LoggedInAndConfigured)
                 {
                     false,                                  // alert
                     false,                                  // playsound
-                    boost::none,                            // custom sound url
+                    std::nullopt,                           // custom sound url
                     std::make_shared<QColor>("#7fe8b7ec"),  // color
                     true,                                   // showInMentions
                 },
@@ -379,7 +378,7 @@ TEST_F(HighlightControllerTest, LoggedInAndConfigured)
                 {
                     true,                                   // alert
                     false,                                  // playsound
-                    boost::none,                            // custom sound url
+                    std::nullopt,                           // custom sound url
                     std::make_shared<QColor>("#7ff19900"),  // color
                     true,                                   // showInMentions
                 },
@@ -399,7 +398,7 @@ TEST_F(HighlightControllerTest, LoggedInAndConfigured)
                 {
                     true,                                   // alert
                     true,                                   // playsound
-                    boost::none,                            // custom sound url
+                    std::nullopt,                           // custom sound url
                     std::make_shared<QColor>("#7f7f3f49"),  // color
                     true,                                   // showInMentions
                 },
@@ -420,7 +419,7 @@ TEST_F(HighlightControllerTest, LoggedInAndConfigured)
                 {
                     false,                                  // alert
                     false,                                  // playsound
-                    boost::none,                            // custom sound url
+                    std::nullopt,                           // custom sound url
                     std::make_shared<QColor>("#6fffffff"),  // color
                     false,
                 },
@@ -442,9 +441,9 @@ TEST_F(HighlightControllerTest, LoggedInAndConfigured)
                 .state = true,  // state
                 .result =
                     {
-                        false,        // alert
-                        false,        // playsound
-                        boost::none,  // custom sound url
+                        false,         // alert
+                        false,         // playsound
+                        std::nullopt,  // custom sound url
                         std::make_shared<QColor>(
                             HighlightPhrase::
                                 FALLBACK_HIGHLIGHT_COLOR),  // color
@@ -468,9 +467,9 @@ TEST_F(HighlightControllerTest, LoggedInAndConfigured)
                 .state = true,  // state
                 .result =
                     {
-                        true,         // alert
-                        true,         // playsound
-                        boost::none,  // custom sound url
+                        true,          // alert
+                        true,          // playsound
+                        std::nullopt,  // custom sound url
                         std::make_shared<QColor>("#7f7f3f49"),  // color
                         true,  // showInMentions
                     },